reorganize geometry code
